share Enum pieces with CoreRT
