+++ src/System.Private.CoreLib/shared/System.Private.CoreLib.Shared.projitems
@@ -222,7 +222,8 @@
     <Compile Include="$(MSBuildThisFileDirectory)System\Environment.SpecialFolder.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)System\Environment.SpecialFolderOption.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)System\EnvironmentVariableTarget.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)System\Enum.cs" Condition="'$(TargetsCoreRT)' != 'true'"/>
+    <Compile Include="$(MSBuildThisFileDirectory)System\Enum.RuntimeType.cs" Condition="'$(TargetsCoreRT)' != 'true'"/>
+    <Compile Include="$(MSBuildThisFileDirectory)System\Enum.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)System\EventArgs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)System\EventHandler.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)System\ExecutionEngineException.cs" />
